+++ src/components/table/user.table.tsx
@@ -81,41 +81,46 @@ const UserTable = (props: Iprops) => {
 
   const [selectedRows, setSelectedRows] = useState<string[]>([]);
   const handleDeleteManyUser = async () => {
-    if (
-      window.confirm(
-        `Are you sure want to delete ${selectedRows.length} users below? `
-      )
-    ) {
-      try {
-        const { data } = await axios.delete(`${process.env.BASE_URL}/user`, {
-          headers: {
-            Authorization: `Bearer ${token}`,
-          },
-          data: {
-            userIds: selectedRows,
-          },
-        });
-        if (data?.status === false) {
-          toast.warning(data?.msg);
+    if (selectedRows.length > 0) {
+      if (
+        window.confirm(
+          `Are you sure want to delete ${selectedRows.length} users below? `
+        )
+      ) {
+        try {
+          const { data } = await axios.delete(`${process.env.BASE_URL}/user`, {
+            headers: {
+              Authorization: `Bearer ${token}`,
+            },
+            data: {
+              userIds: selectedRows,
+            },
+          });
+          if (data?.status === false) {
+            toast.warning(data?.msg);
+            return;
+          }
+          if (data?.status === true) {
+            toast.success(data?.msg);
+            const res = await axios.get(
+              `${process.env.BASE_URL}/user?s=${keyWordSearch}&page=${currentPage}&limit=${itemsPerPage}`,
+              {
+                headers: {
+                  Authorization: `Bearer ${token}`,
+                },
+              }
+            );
+            setRecords(res?.data?.data);
+            setAllRows(res.data?.totalUsers);
+          }
+        } catch (error: any) {
+          toast.error(error?.response?.data?.message);
           return;
         }
-        if (data?.status === true) {
-          toast.success(data?.msg);
-          const res = await axios.get(
-            `${process.env.BASE_URL}/user?s=${keyWordSearch}&page=${currentPage}&limit=${itemsPerPage}`,
-            {
-              headers: {
-                Authorization: `Bearer ${token}`,
-              },
-            }
-          );
-          setRecords(res?.data?.data);
-          setAllRows(res.data?.totalUsers);
-        }
-      } catch (error: any) {
-        toast.error(error?.response?.data?.message);
-        return;
       }
+    } else {
+      toast.warning("Please pick users to delete them");
+      return;
     }
   };
